Attempted to create a button for the inventory items
The button still won't work as it functions on conditional logic on whether the item is already part of an order (in which case there should be no button) or if the item is in the user's cart (in which case a "remove from cart" button would be displayed). If it's neither of those cases then it should be an item that can be added to the user's cart, and an "add to cart" button should be displayed.